Fix navigation text Field
